Move unit tests into their own directory; Organize jest config
