+++ package.json
@@ -3,27 +3,33 @@
   "version": "4.9.45",
   "description": "Dark mode for every website",
   "scripts": {
-    "api": "node tasks/build.js --api",
     "benchmark-server": "node tests/benchmark-server/index.js",
+    "api": "node tasks/build.js --api",
     "build": "node tasks/build.js --release",
-    "code-style": "eslint --ignore-pattern \"!.eslintplugin.js\" --cache --fix \"src/**/*.ts\" \"src/**/*.tsx\" \"tasks/**/*.js\" \"tests/[!coverage]**/*.js\" \"tests/**/*.ts\" \".eslintrc.js\" \"index.d.ts\" \".eslintplugin.js\"",
+    "build:all": "node tasks/build.js --debug --release --api",
     "debug": "node tasks/build.js --debug",
     "debug:watch": "node tasks/build.js --debug --watch",
-    "lint": "eslint --ignore-pattern \"!.eslintplugin.js\" \"src/**/*.ts\" \"src/**/*.tsx\" \"tasks/**/*.js\" \"tests/**/*.ts\" \"tests/[!coverage]**/*.js\" \"index.d.ts\" \".eslintplugin.js\"",
     "prepublishOnly": "npm test && npm run api",
     "release": "npm test && npm run lint && node tasks/build.js --release",
-    "test": "jest --config=tests/jest.config.js",
-    "test:browser": "npm run debug && jest --config=tests/browser/jest.config.js --runInBand",
-    "test:chrome": "npm run debug && jest --config=tests/browser/jest.config.chrome.js --runInBand",
-    "test:ci": "jest --config=tests/jest.config.js --runInBand",
-    "test:config": "jest --config=tests/config/jest.config.js",
-    "test:config:debug": "node --inspect-brk ./node_modules/jest/bin/jest --config=tests/config/jest.config.js --runInBand --no-cache --watch",
-    "test:coverage": "jest --config=tests/jest.config.js --coverage",
-    "test:firefox": "npm run debug && jest --config=tests/browser/jest.config.firefox.js --runInBand",
+    "code-style": "eslint --ignore-pattern \"!.eslintplugin.js\" --cache --fix \"src/**/*.ts\" \"src/**/*.tsx\" \"tasks/**/*.js\" \"tests/[!coverage]**/*.js\" \"tests/**/*.ts\" \".eslintrc.js\" \"index.d.ts\" \".eslintplugin.js\"",
+    "lint": "eslint --ignore-pattern \"!.eslintplugin.js\" \"src/**/*.ts\" \"src/**/*.tsx\" \"tasks/**/*.js\" \"tests/**/*.ts\" \"tests/[!coverage]**/*.js\" \"index.d.ts\" \".eslintplugin.js\"",
+
+    "test": "npm run test:unit",
+    "test:ci": "npm run test:unit",
+    "test:all": "npm run test:unit; npm run test:browser; npm run test:inject; npm run test:project",
+
+    "test:unit": "jest --config=tests/unit/jest.config.js",
+    "test:unit:debug": "node --inspect-brk ./node_modules/jest/bin/jest --config=tests/unit/jest.config.js --runInBand --no-cache --watch",
+    "test:coverage": "jest --config=tests/unit/jest.config.js --coverage",
+
+    "test:browser": "npm run debug && npm run test:chrome && npm run test:firefox",
+    "test:chrome": "TEST_BROWSER=chrome jest --config=tests/browser/jest.config.js --runInBand",
+    "test:firefox": "TEST_BROWSER=firefox jest --config=tests/browser/jest.config.js --runInBand",
+
     "test:inject": "node tests/inject/run.js",
     "test:inject:debug": "node tests/inject/run.js --debug",
-    "test:utils": "jest --config=tests/utils/jest.config.js",
-    "test:utils:debug": "node --inspect-brk ./node_modules/jest/bin/jest --config=tests/utils/jest.config.js --runInBand --no-cache --watch"
+
+    "test:project": "jest --config=tests/project/jest.config.js"
   },
   "main": "darkreader.js",
   "repository": {
@@ -68,13 +74,14 @@
     "@rollup/plugin-replace": "3.0.1",
     "@rollup/plugin-typescript": "8.3.0",
     "@rollup/pluginutils": "4.1.2",
-    "@types/chrome": "0.0.179",
+    "@types/chrome": "0.0.176",
+    "@types/eslint": "8.4.1",
     "@types/jasmine": "3.10.3",
     "@types/jest": "27.4.0",
     "@types/karma": "6.3.2",
-    "@types/offscreencanvas": "2019.6.4",
+    "@types/karma-coverage": "^2.0.1",
     "@types/node": "17.0.9",
-    "@types/filesystem": "0.0.32",
+    "@types/offscreencanvas": "2019.6.4",
     "@typescript-eslint/eslint-plugin": "5.9.1",
     "@typescript-eslint/parser": "5.9.1",
     "chokidar": "3.5.2",
