+++ tests/browser/e2e/devtools.tests.ts
@@ -1,4 +1,4 @@
-import {multiline, timeout} from '../../test-utils';
+import {multiline, timeout} from '../../support/test-utils';
 
 async function loadBasicPage() {
     await loadTestPage({
