--- tests/browser/jest.config.chrome.js
@@ -1,9 +1,0 @@
-const jestConfig = require('./jest.config.shared');
-
-module.exports = {
-    ...jestConfig,
-    globals: {
-        ...jestConfig.globals,
-        product: 'chrome',
-    },
-};
