--- tests/generators/utils/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../../../src/tsconfig",
-    "compilerOptions": {
-        "module": "commonjs",
-        "outDir": "",
-        "types": [
-            "jest",
-            "node"
-        ]
-    }
-}
