--- tests/inject/jest.config.js
@@ -1,23 +1,0 @@
-module.exports = {
-    verbose: true,
-    transform: {
-        '^.+\\.ts$': 'ts-jest'
-    },
-    testRegex: 'tests/inject/.*\\.tests\\.ts$',
-    moduleFileExtensions: [
-        'ts',
-        'js'
-    ],
-    rootDir: '../../',
-    collectCoverage: false,
-    coverageDirectory: 'tests/coverage',
-    collectCoverageFrom: [
-        '<rootDir>/src/**/*.ts',
-    ],
-    globals: {
-        'ts-jest': {
-            tsconfig: './tests/inject/tsconfig.json'
-        },
-        __DEBUG__: false,
-    }
-};
