+++ tests/inject/tsconfig.json
@@ -1,10 +1,29 @@
 {
-    "extends": "../../src/tsconfig",
+    "$schema": "https://json.schemastore.org/tsconfig",
     "compilerOptions": {
+        "target": "ES2019",
+        "module": "ES2015",
+        "lib": [
+            "ES2015",
+            "ES2017",
+            "DOM",
+            "DOM.Iterable"
+        ],
         "types": [
             "chrome",
-            "jasmine",
-            "offscreencanvas"
-        ]
-    }
+            "offscreencanvas",
+            "jasmine"
+        ],
+        "allowJs": true,
+        "downlevelIteration": true,
+        "esModuleInterop": true,
+        "jsx": "react",
+        "jsxFactory": "m",
+        "moduleResolution": "Node",
+        "noEmit": true,
+        "noImplicitAny": true
+    },
+    "exclude": [
+        "./coverage",
+    ]
 }
