--- tests/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    projects: [
-        'tests/config/jest.config.js',
-        'tests/utils/jest.config.js',
-        'tests/generators/utils/jest.config.js',
-    ],
-    verbose: true,
-};
