+++ tests/unit/config/locales.tests.ts
@@ -1,9 +1,9 @@
 import {readFile as fsReadFile, readdir as fsReadDir} from 'fs';
-import {resolve as resolvePath} from 'path';
+import {rootPath} from '../../support/test-utils';
 
 function readDir(dir: string) {
     return new Promise<string[]>((resolve, reject) => {
-        fsReadDir(resolvePath(__dirname, dir), (err, files) => {
+        fsReadDir(dir, (err, files) => {
             if (err) {
                 reject(err);
                 return;
@@ -15,7 +15,7 @@ function readDir(dir: string) {
 
 function readLocale(name: string) {
     return new Promise<string>((resolve, reject) => {
-        fsReadFile(resolvePath(__dirname, '../../src/_locales/', name), {encoding: 'utf-8'}, (err, data) => {
+        fsReadFile(rootPath('src/_locales', name), {encoding: 'utf-8'}, (err, data) => {
             if (err) {
                 reject(err);
                 return;
@@ -26,7 +26,7 @@ function readLocale(name: string) {
 }
 
 test('Locales', async () => {
-    const files = await readDir('../../src/_locales');
+    const files = await readDir(rootPath('src/_locales'));
     const enLocale = await readLocale('en.config');
     const enLines = enLocale.split('\n');
     const locales: string[] = [];
