+++ tests/unit/utils/color.tests.ts
@@ -1,5 +1,5 @@
-import type {HSLA} from '../../src/utils/color';
-import {lowerCalcExpression, parse, hslToRGB, rgbToHSL, rgbToString, rgbToHexString, hslToString} from '../../src/utils/color';
+import type {HSLA} from '../../../src/utils/color';
+import {lowerCalcExpression, parse, hslToRGB, rgbToHSL, rgbToString, rgbToHexString, hslToString} from '../../../src/utils/color';
 
 test('Color parsing', () => {
     expect(parse('rgb(255,0,153)')).toEqual({r: 255, g: 0, b: 153, a: 1});
