+++ tests/unit/utils/text.tests.ts
@@ -1,4 +1,4 @@
-import {formatCSS, getParenthesesRange} from '../../src/utils/text';
+import {formatCSS, getParenthesesRange} from '../../../src/utils/text';
 
 test('CSS formatting', () => {
     expect(formatCSS('div { color: red; }'))
